requirements.txt
kafka-python
python-schema-registry-client
